Decompose game list row components
Refs #158
